added relation between user and data

+++ laragigs/routes/web.php
@@ -1,10 +1,10 @@
 <?php
 
-use App\Models\Listing;
+use App\Http\Controllers\ListingController;
+use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\UserController;
-use App\Http\Controllers\ListingController;
+use App\Models\Listing;
 
 /*
 |--------------------------------------------------------------------------
@@ -12,60 +12,55 @@ use App\Http\Controllers\ListingController;
 |--------------------------------------------------------------------------
 |
 | Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider and all of them will
-| be assigned to the "web" middleware group. Make something great!
+| routes are loaded by the RouteServiceProvider within a group which
+| contains the "web" middleware group. Now create something great!
 |
 */
 
+// Common Resource Routes:
+// index - Show all listings
+// show - Show single listing
+// create - Show form to create new listing
+// store - Store new listing
+// edit - Show form to edit listing
+// update - Update listing
+// destroy - Delete listing  
 
-
-
-// Common resources route 
-// Verb      | URI                        | Action       | Route Name
-// --------------------------------------------------------------------------
-// GET       | /resource-name             | index        | resource-name.index
-// GET       | /resource-name/create      | create       | resource-name.create
-// POST      | /resource-name             | store        | resource-name.store
-// GET       | /resource-name/{id}        | show         | resource-name.show
-// GET       | /resource-name/{id}/edit   | edit         | resource-name.edit
-// PUT/PATCH | /resource-name/{id}        | update       | resource-name.update
-// DELETE    | /resource-name/{id}        | destroy      | resource-name.destroy
-
-//All Listing
+// All Listings
 Route::get('/', [ListingController::class, 'index']);
 
+// Show Create Form
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
-//Show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+// Store Listing Data
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
-//Store listing data
-Route::post('/listings', [ListingController::class, 'store']);
+// Show Edit Form
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
-//Show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+// Update Listing
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
-// Update listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+// Delete Listing
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
-// Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
-// Single listing
+// Single Listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
+// Show Register/Create Form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
-//Show Register/Create Form
-Route::get('/register', [UserController::class, 'create']);
-
-//Create new user
+// Create New User
 Route::post('/users', [UserController::class, 'store']);
 
-//Log user out
-Route::post('/logout', [UserController::class, 'logout']);
+// Log User Out
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
+// Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
-//Sow Login form
-Route::get('/login', [UserController::class, 'login']);
-
-//Login User
+// Log In User
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
